% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/econ_growth.R
\docType{data}
\name{econ_growth}
\alias{econ_growth}
\title{Economic Growth Data for 15 Industrialized Democracies, 1974-1980}
\format{A data frame with 15 rows and 4 columns.
\tabular{lll}{
  \code{country}     \tab \code{character}        \tab Country name \cr
  \code{econ_growth} \tab \code{double}           \tab Economic Growth \cr
  \code{labor_org}   \tab \code{double}           \tab Labor organization \cr
  \code{socia_dem}   \tab \code{social_democracy} \tab Social democracy
}}
\source{
Western, B. (1995) \dQuote{\href{http://www.jstor.org/stable/2111654}{Concepts and Suggestions for Robust Regression Analysis}.}
  \emph{American Journal of Political Science}.
}
\usage{
econ_growth
}
\description{
Lange and Garret (1985) economic growth data for 15 industrialized
democracies.
}
\references{
Lange, P., Garrett, G. (1985) \dQuote{The Politics of Growth: Strategic Interaction and Economic Performance in the Advanced Industrial Democracies, 1974-1980.}
  \emph{Journal of Politics}.
}
\keyword{datasets}
