% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eightschools.R
\docType{data}
\name{eightschools}
\alias{eightschools}
\title{Eight Schools}
\format{A data frame with eight rows and two columns,
\tabular{lll}{
\code{y} \tab \code{numeric} \tab The mean change in test scores \cr
\code{sigma} \tab \code{numeric} \tab The standard deviation of change in test scores
}}
\source{
\url{https://github.com/stan-dev/rstan/wiki/RStan-Getting-Started}
}
\usage{
eightschools
}
\description{
Data on the results of coaching effects for eight schools.
}
\references{
\emph{Bayesian Data Analysis}. 3rd edition.
}
\keyword{datasets}
