% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_stan_model}
\alias{get_stan_model}
\title{Get and compile an included Stan model}
\usage{
get_stan_model(modelname, ...)
}
\arguments{
\item{modelname}{The name of the model to get.}

\item{...}{Passed to \code{\link[rstan]{stan_model}}}
}
\value{
An object of class \code{"\linkS4class{stanmodel}"}.
}
\description{
Get and compile an Stan model included in the package.
}
\seealso{
Other models: \code{\link{list_stan_models}}
}
