% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jrnold.bayes.notes-package.R
\docType{package}
\name{jrnold.bayes.notes-package}
\alias{jrnold.bayes.notes}
\alias{jrnold.bayes.notes-package}
\title{jrnold.bayes.notes: Miscellaneous Functions and Data for Bayesian Notes}
\description{
Package to install dependencies, Stan models,
and data for <https://github.com/jrnold/bayesian_notes>.
}
\author{
\strong{Maintainer}: Jeffrey Arnold \email{jeffrey.arnold@gmail.com}

}
\keyword{internal}
