% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_stan_models}
\alias{list_stan_models}
\title{List included Stan models}
\usage{
list_stan_models()
}
\value{
A character vector with the names of the all
  the models.
}
\description{
List the Stan models included in the package.
}
\seealso{
Other models: \code{\link{get_stan_model}}
}
