% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/politics_and_need.R
\docType{data}
\name{politics_and_need}
\alias{politics_and_need}
\title{Governors' Decisions to Support or Oppose the 2010 Medicaid Expansions}
\format{A data frame with 50 rows and 10 variables.}
\source{
\url{https://github.com/carlislerainey/separation/blob/cbe6e1fa12ff4507465e98fbe58740ae928bc9da/data/politics_and_need.rda}
}
\usage{
politics_and_need
}
\description{
Data on governors' decisions to support or oppose Medicaid expansions
offered under the 2010 Patient Protection and Affordable Care Act
used in Barrilleaux and Rainey (2014).
}
\references{
Barrilleaux, C., Rainey, C. (2014) \dQuote{The Politics of Need Examining Governors’ Decisions to Oppose the \sQuote{Obamacare} Medicaid Expansion.} \emph{State Politics and Policy Quarterly}
\href{https://dx.doi.org/10.1177/1532440014561644}{DOI:10.1177/1532440014561644}

Rainey, C. (2016) \dQuote{Dealing with Separation in Logistic Regression Models}
\emph{Political Analysis}. \href{https://doi.org/10.1093/pan/mpw014}{doi:10.1093/pan/mpw014}
}
\keyword{datasets}
