% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rats.R
\docType{data}
\name{rats}
\alias{rats}
\title{Rat Weights}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 180 rows and 2 columns.}
\source{
\url{https://raw.github.com/wiki/stan-dev/rstan/rats.txt}
}
\usage{
rats
}
\description{
Data on 30 young rats whose weights were measured weekly for five weeks
used in Gelfand et al. (1990). This data can be modeled using a hierarchical
normal model.
}
\references{
Gelfand, A. E., Hills S. E., Racine-Poon, A., and Smith A. F. M. (1990).
\dQuote{Illustration of Bayesian Inference in Normal Data Models Using Gibbs Sampling},
\emph{Journal of the American Statistical Association}, 85, 972-985.

OpenBUGS example, \href{http://www.openbugs.net/Examples/Rats.html}{Rats: a normal hierarchical model}
}
\keyword{datasets}
