% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/separation.R
\docType{data}
\name{separation}
\alias{separation}
\alias{CompleteSeparation}
\alias{QuasiSeparation}
\title{Fake data illustrating separation}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 8 rows and 3 columns.}
\usage{
CompleteSeparation

QuasiSeparation
}
\description{
Fake datasets illustrating complete separation
(\code{CompleteSeparation}) and quasi-separation
(\code{QuasiSeparation})
}
\keyword{datasets}
