% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votechoice.R
\docType{data}
\name{votechoice}
\alias{votechoice}
\title{Vote Choice in the 2004 US Presidential Election}
\format{A data frame with 1,000 rows and 12 columns.
\tabular{lll}{
\code{caseid} \tab \code{integer} \tab Survey identifier \cr
\code{ideol7b} \tab \code{factor} \tab Ideology \cr
\code{retecon} \tab \code{factor} \tab Retrospective economic evaluations \cr
\code{partyid} \tab \code{factor} \tab Party identification \cr
\code{bushiraq} \tab \code{factor} \tab Approval of Handling of Iraq War \cr
\code{white} \tab \code{logical} \tab Is White? \cr
\code{female} \tab \code{logical} \tab Is Female? \cr
\code{education} \tab \code{factor} \tab Education level \cr
\code{income} \tab \code{factor} \tab 1-23 where 1 = less than \$2,999 and 23 = \$120,000+ \cr
\code{bushvote} \tab \code{logical} \tab Voted for Bush?
}}
\source{
Hanmer, M. J.,  Kalkan, K. O. (2012)
 \dQuote{Replication data for: Behind the Curve: Clarifying the Best Approach to Calculating Predicted Probabilities and Marginal Effects from Limited Dependent Variable Models.}
 \url{https://hdl.handle.net/1902.1/18542},
 Harvard Dataverse, V2,
 \verb{UNF:5:m5jfkzRK4z7smtwB6ijZxQ==}.
}
\usage{
votechoice
}
\description{
A random sample of 383 observations from the 2004 National Election Survey
with data on vote choice in the 2004 U.S. presidential election
(George W. Bush or John Kerry).
This is used as an example in Hanmer and Kalkan (2004).
}
\references{
Hanmer, M. J., Kalkan, K. O. (2013)
  \dQuote{Behind the Curve: Clarifying the Best Approach to Calculating Predicted Probabilities and Marginal Effects from Limited Dependent Variable Models.}
  \emph{American Journal of Political Science}.
  \href{https://dx.doi.org/10.1111/j.1540-5907.2012.00602.x http://dx.doi.org/10.1111/j.1540-5907.2012.00602.x}{doi:10.1111/j.1540-5907.2012.00602.x}.
}
\keyword{datasets}
